AI1/requirements.txt
flask
flask-cors
flask-bcrypt
flask-login
openai
requests          # ElevenLabs
python-dotenv     # 读取 .env
